% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generics.R, R/paga.R
\name{CalculatePAGA}
\alias{CalculatePAGA}
\alias{CalculatePAGA.default}
\alias{CalculatePAGA.Seurat}
\title{Calculate PAGA connectivities}
\usage{
CalculatePAGA(object, ...)

\method{CalculatePAGA}{default}(object, group)

\method{CalculatePAGA}{Seurat}(object, group.by = NULL, assay = NULL, graph.name = NULL)
}
\arguments{
\item{object}{An object}

\item{...}{Arguments passed to other methods}

\item{group}{A vector of the groups of each cell that specifies
the mapping. Its elements correspond to the KNN matrix.}

\item{group.by}{Name of one or more metadata columns to group
cells by (for example, orig.ident)}

\item{assay}{Assay used in calculate KNN matrix}

\item{graph.name}{Name of graph to use for the paga algorithm}
}
\value{
This function can return a adjacency matrix with the paga
connectivities information of given cell groups . When running on
a \code{\link{Seurat}} object, this function returns the
\code{\link{Seurat}} object with the Graphs stored in their respective
slots. Names of the Graph can be found with \code{\link{Graphs}}.
}
\description{
Computes the PAGA connectivities for a given KNN adjacency matrix of cells.
Calculateconnectivities between every cell groups and its nearest neighbors.
}
\examples{
library(igraph)
pbmc<-CalculatePAGA(object=pbmc, group.by='celltype')
plot(graph_from_adjacency_matrix(pbmc@graphs$RNA_paga))

}
