% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{PlotPAGA}
\alias{PlotPAGA}
\title{Plot PAGA with cell embedding}
\usage{
PlotPAGA(
  object,
  group.by = NULL,
  reduction = NULL,
  dims = c(1, 2),
  threshold = 0.01,
  cell.plot = F,
  assay = NULL,
  slot = "data",
  feature = NULL,
  min.cutoff = NULL,
  max.cutoff = NULL,
  feature.cols = c("grey", "red"),
  pt.size = 1,
  group.cols = NULL,
  label = F,
  label.size = 5,
  raster.dpi = 500
)
}
\arguments{
\item{object}{Seurat object}

\item{group.by}{Name of one or more metadata columns to group (color) cells by
(for example, orig.ident); pass 'ident' to group by identity class}

\item{reduction}{Which dimensionality reduction to use. If not specified, first searches for umap, then tsne, then pca}

\item{dims}{Dimensions to plot, must be a two-length numeric vector specifying x- and y-dimensions}

\item{threshold}{Threshold for connectivities between cell groups, default is 0.01}

\item{cell.plot}{Weature to plot cell points, default is not.}

\item{assay}{Assay used to calculate cell group connectivities}

\item{slot}{Which slot to pull expression data from? Default is data slot}

\item{feature}{character of a feature to plot. Feature can come from:
\itemize{
    \item An \code{Assay} feature (e.g. a gene name - "MS4A1")
    \item A column name from meta.data (e.g. mitochondrial percentage - "percent.mito")
    \item A column name from a \code{DimReduc} object corresponding to the cell embedding values
    (e.g. the PC 1 scores - "PC_1")
}}

\item{min.cutoff, max.cutoff}{Vector of minimum and maximum cutoff values for feature}

\item{feature.cols}{The multiple colors to form the gradient over. Provide as string vector with
the first color corresponding to low values, the second to high.}

\item{pt.size}{Adjust cell point size for plotting, default is 1.}

\item{group.cols}{Vector of colors, each color corresponds to an identity class. By default, hue_pal is used.}

\item{label}{Whether to label the clusters}

\item{label.size}{Sets size of labels}

\item{raster.dpi}{Cells convert points to raster format, Pixel resolution for rasterized plot.}
}
\value{
A ggplot object with PAGA and a feature in reduction.
}
\description{
Plot the PAGA connectivities from connectivities matrix of cell groups.
If parameter show.cell=F, only shows cell groups, each cell group is a
point, point size indicate cell count in group, point position is the
center of the position of cells based on the cell embeddings determined
by the reduction technique. otherwise a 2D scatter plot where each point
is a cell with same embeddings with cell groups will be put under cell
group points. Cells are colored by their group (can be changed with the
feature parameter).
}
\examples{
PlotPAGA(object=pbmc)
PlotPAGA(object=pbmc, feature='PTPRC')


}
